Maintain-structure 'code updated'
