+++ src/app/model/freeshareDashboard.ts
@@ -1,0 +1,8 @@
+export class freeshareDashboardModule { 
+    address:string;
+    dob:string;
+    id:any;
+    locked:any;
+    name:string;
+    phone:string;
+  }
