+++ src/app/model/product.ts
@@ -1,0 +1,5 @@
+export class Product { 
+    ProductName:string;
+    Deleted:Number;
+    ProductId:Number;
+  }
