--- src/app/module/freeshareDashboard.module.ts
@@ -1,18 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule
-  ]
-})
-
-export class freeshareDashboardModule { 
-  address:string;
-  dob:string;
-  id:any;
-  locked:any;
-  name:string;
-  phone:string;
-}
