--- src/app/module/product.module.ts
@@ -1,14 +1,0 @@
-import { NgModule } from '@angular/core';
-import { CommonModule } from '@angular/common';
-
-@NgModule({
-  declarations: [],
-  imports: [
-    CommonModule
-  ]
-})
-export class ProductModule { 
-  ProductName:string;
-  Deleted:Number;
-  ProductId:Number;
-}
