+++ src/app/service/freeshare-dashboard.service.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@angular/core';
 import {HttpClient,HttpHeaders} from '@angular/common/http';
-import { ProductModule } from '../module/product.module';
+import { Product } from '../model/product';
 
 @Injectable({
   providedIn: 'root'
@@ -20,18 +20,18 @@ export class freeshareDashboardService {
   
   getAllProduct()
     {
-        return this.http.get<ProductModule[]>(this.baseUrl+"GetAllProductDetails");
+        return this.http.get<Product[]>(this.baseUrl+"GetAllProductDetails");
     }
 
-    insertProduct(productName:ProductModule){
+    insertProduct(productName:Product){
       return this.http.post(this.baseUrl+"InsertProductDetails",JSON.stringify(productName),this.httpOptions);
     }
 
     getProductById(id:number){
-      return this.http.get<ProductModule[]>(this.baseUrl+"GetProductById/"+id);
+      return this.http.get<Product[]>(this.baseUrl+"GetProductById/"+id);
     }
 
-    updateProduct(productName:ProductModule){
+    updateProduct(productName:Product){
       return this.http.put(this.baseUrl+"UpdateProductDetails",JSON.stringify(productName),this.httpOptions);
     }
     deleteProduct(id:number){
